Rename kslack -> arg. Add arg to all scripts and remove bots. (@Schalk1e)
