+++ rapidpro/helpers/arg.py
@@ -1,8 +1,5 @@
 import os
 
-DS_SPAM = os.environ["DS_SPAM_WEBHOOK_URL"]
-DS_ERRORS = os.environ["DS_ERRORS_WEBHOOK_URL"]
-
 import requests
 import argparse
 
@@ -15,18 +12,3 @@ command_line_parser.add_argument("-id", "--org_id", help="Rapidpro organisation 
 command_line_parser.add_argument("-k", "--api_key", help="Rapidpro API key value (string).", required=True)
 command_line_parser.add_argument("-u", "--base_url", help="Rapidpro base URL value (string).", required=True)
 command_line_args = command_line_parser.parse_args()
-
-def post_message(message,channel):  
-    if command_line_args.test_mode or command_line_args.temp_channel:
-        channel = command_line_args.temp_channel or command_line_args.test_mode and 'ds-testing'
-        print(f'Posting to slack channel {channel}:\n{message}')
-    headers = {
-        'Content-type': 'application/json',
-    }
-    data = str({ "text" : "{0}".format(message) })
-    if channel == 'ds-spam':
-        url = DS_SPAM  
-    if channel == 'ds-errors':
-        url = DS_ERRORS
-        
-    response = requests.post(url, headers=headers, data=data)
