+++ rapidpro/rpp_ftbl_contacts.py
@@ -1,7 +1,7 @@
 import os
 
 from click import command
-from helpers.kslack import command_line_args, post_message
+from helpers.arg import command_line_args
 
 os.environ["API_KEY"] = command_line_args.api_key
 os.environ["BASE_URL"] = command_line_args.base_url
@@ -33,10 +33,7 @@ if __name__ == '__main__':
 			warehouse.update(file_name="SQL/Conflicts/rpp_ftbl_contacts_conflict.sql")
 			warehouse.update(file_name='SQL/Analytic/rpp_ftbl_contacts_update.sql')
 			warehouse.drop('staging_rpp_ftbl_contact')
-			post_message(message=f'rpp_ftbl_contact table ran successfull. {contacts.shape[0]} rows updated', channel="ds-spam")
-		else:
-			post_message(message=f'rpp_ftbl_contact table ran successfull. No rows updated', channel="ds-spam")
+
 	except Exception as e:
-		post_message(message=f'rpp_ftbl_contact.py failed: {e}', channel="ds-errors")
 		raise
