+++ rapidpro/rpp_ftbl_flows_flow.py
@@ -1,7 +1,7 @@
 import os
 
 from click import command
-from helpers.kslack import command_line_args, post_message
+from helpers.arg import command_line_args
 
 os.environ["API_KEY"] = command_line_args.api_key
 os.environ["BASE_URL"] = command_line_args.base_url
@@ -33,9 +33,5 @@ if __name__ == '__main__':
 			warehouse.update(file_name='SQL/Analytic/rpp_ftbl_flows_flow_update.sql')
 			warehouse.drop('staging_rpp_ftbl_flows_flow')
 
-			post_message(message=f'rpp_ftbl_flows_flow table ran successfull. {flows_flow.shape[0]} rows updated', channel="ds-spam")
-		else:
-			post_message(message=f'rpp_ftbl_flows_flow table ran successfull. No rows updated', channel="ds-spam")
 	except Exception as e:
-		post_message(message=f'rpp_ftbl_flows_flow.py failed: {e}', channel="ds-errors")
 		raise
