+++ rapidpro/rpp_ftbl_msgs_broadcast_urns.py
@@ -1,7 +1,7 @@
 import os
 
 from click import command
-from helpers.kslack import command_line_args, post_message
+from helpers.arg import command_line_args
 
 os.environ["API_KEY"] = command_line_args.api_key
 os.environ["BASE_URL"] = command_line_args.base_url
@@ -34,9 +34,5 @@ if __name__ == '__main__':
 			warehouse.update(file_name='SQL/Analytic/rpp_ftbl_msgs_broadcast_urns_update.sql')
 			warehouse.drop('staging_rpp_ftbl_msgs_broadcast_urns')
 
-			post_message(message=f'rpp_ftbl_msgs_broadcast_urns table ran successfull. {msgs_broadcast_urns.shape[0]} rows updated', channel="ds-spam")
-		else:
-			post_message(message=f'rpp_ftbl_msgs_broadcast_urns table ran successfull. No rows updated', channel="ds-spam")
 	except Exception as e:
-		post_message(message=f'rpp_ftbl_msgs_broadcast_urns.py failed: {e}', channel="ds-errors")
 		raise
